While importing toolset.xml, only warn about missing required tools.
We print a warning when the class of a required tool is not found and
continue with the next tool.  The previous behaviour could break the
install or uninstall of any add-on, as the missing class may easily be
from a different unrelated add-on that is no longer available in the
zope instance.

Conflicts:
	Products/GenericSetup/CHANGES.txt
	Products/GenericSetup/tool.py